requirements.txt
django
djangorestframework-simplejwt
djangorestframework
django-cors-headers
torch
transformers
python-docx
pdfminer.six